Move upload functionality into MyObservations, out of ObsEditProvider
