+++ src/components/MyObservations/Header.js
@@ -21,8 +21,10 @@ type Props = {
   currentUser: ?Object,
   hideToolbar: boolean,
   setHeightAboveToolbar: Function,
-  allObsToUpload: Array<Object>,
-  setShowLoginSheet: Function
+  uploadState: Object,
+  uploadMultipleObservations: Function,
+  stopUpload: Function,
+  syncObservations: Function
 }
 
 const Header = ( {
@@ -31,8 +33,10 @@ const Header = ( {
   currentUser,
   hideToolbar,
   setHeightAboveToolbar,
-  allObsToUpload,
-  setShowLoginSheet
+  uploadState,
+  uploadMultipleObservations,
+  stopUpload,
+  syncObservations
 }: Props ): Node => {
   const theme = useTheme( );
   const navigation = useNavigation( );
@@ -119,8 +123,10 @@ const Header = ( {
           toggleLayout={toggleLayout}
           layout={layout}
           numUnuploadedObs={numUnuploadedObs}
-          allObsToUpload={allObsToUpload}
-          setShowLoginSheet={setShowLoginSheet}
+          uploadState={uploadState}
+          uploadMultipleObservations={uploadMultipleObservations}
+          stopUpload={stopUpload}
+          syncObservations={syncObservations}
         />
       )}
     </>
