+++ src/components/SharedComponents/ObservationsFlashList/ObsItem.js
@@ -10,13 +10,14 @@ type Props = {
   observation: Object,
   gridItemWidth: number,
   layout: "list" | "grid",
-  setShowLoginSheet: Function
+  uploadObservation?: Function,
+  uploadState: Object
 };
 
 const GUTTER = 15;
 
 const ObsItem = ( {
-  observation, layout, gridItemWidth, setShowLoginSheet
+  observation, layout, gridItemWidth, uploadState, uploadObservation
 }: Props ): Node => (
   <MyObservationsPressable observation={observation} testID="MyObservationsPressable">
     {
@@ -31,13 +32,15 @@ const ObsItem = ( {
               width: gridItemWidth,
               margin: GUTTER / 2
             }}
-            setShowLoginSheet={setShowLoginSheet}
+            uploadObservation={uploadObservation}
+            uploadState={uploadState}
           />
         )
         : (
           <ObsListItem
             observation={observation}
-            setShowLoginSheet={setShowLoginSheet}
+            uploadObservation={uploadObservation}
+            uploadState={uploadState}
           />
         )
     }
